inject logic into a custom component
