+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:paddingTop="?attr/actionBarSize">
 
-    <com.google.android.material.bottomnavigation.BottomNavigationView
+    <com.erictel.bottomdynamic.BottomNavDynamic
         android:id="@+id/nav_view"
         style="@style/BottomNavigationIconStyle2"
         android:layout_width="match_parent"
